@@ -4,29 +4,29 @@ archivesBaseName = 'loom'
 
 allprojects {
     apply plugin: 'checkstyle'
-    apply plugin: 'findbugs'
-    apply plugin: 'pmd'
+//    apply plugin: 'findbugs'
+//    apply plugin: 'pmd'
 
     group = 'builders.loom'
     version = '1.0.0'
 
-    sourceCompatibility = 1.8
+    sourceCompatibility = 9
 
     checkstyle {
         configFile = file("${rootDir}/config/checkstyle/checkstyle.xml")
         configProperties = [samedir: "${rootDir}/config/checkstyle", project_loc: "${rootDir}"]
     }
 
-    pmd {
-        consoleOutput = true
-    }
-
-    tasks.withType(FindBugs) {
-        reports {
-            xml.enabled = false
-            html.enabled = true
-        }
-    }
+//    pmd {
+//        consoleOutput = true
+//    }
+//
+//    tasks.withType(FindBugs) {
+//        reports {
+//            xml.enabled = false
+//            html.enabled = true
+//        }
+//    }
 
     repositories {
         jcenter()
@@ -112,7 +112,6 @@ task zip(type: Zip) {
             from project(':plugin-eclipse').configurations.runtime
         }
         into("scripts") {
-            from 'scripts/build.yml'
             from 'scripts/loom'
             from 'scripts/loom.cmd'
         }
