@@ -20,6 +20,6 @@ public interface ProductRepository {
 
     ProductPromise lookup(String productId);
 
-    void createProduct(String productId);
+    void createProduct(final String moduleName, String productId);
 
 }
