@@ -21,9 +21,11 @@ import java.util.function.Supplier;
 
 public interface TaskRegistry {
 
+    @SuppressWarnings("checkstyle:parameternumber")
     void registerTask(String pluginName, String taskName, Supplier<Task> taskSupplier,
                       String providedProduct, boolean intermediateProduct,
-                      Set<String> usedProducts, String description);
+                      Set<String> usedProducts, Set<String> importedProducts,
+                      Set<String> importedAllProducts, String description);
 
     void registerGoal(String pluginName, String goalName, Set<String> usedProducts);
 
