@@ -21,7 +21,7 @@ import java.util.Optional;
 public interface Product {
 
     /**
-     * Spring Boot application: loombuild/springboot/springboot-1.0.0-SNAPSHOT-fatjar.jar.
+     * Spring Boot application: build/springboot/springboot-1.0.0-SNAPSHOT-fatjar.jar.
      */
     Optional<String> outputInfo();
 
