@@ -17,9 +17,7 @@
 package builders.loom.plugin.java;
 
 import builders.loom.api.AbstractPlugin;
-import builders.loom.api.BuildConfig;
 import builders.loom.api.CompileTarget;
-import builders.loom.api.RuntimeConfiguration;
 
 @SuppressWarnings("checkstyle:classdataabstractioncoupling")
 public class JavaPlugin extends AbstractPlugin<JavaPluginSettings> {
@@ -30,65 +28,62 @@ public class JavaPlugin extends AbstractPlugin<JavaPluginSettings> {
 
     @Override
     public void configure() {
-        final BuildConfig buildConfig = getBuildConfig();
-        final RuntimeConfiguration runtimeConfiguration = getRuntimeConfiguration();
-
         task("provideSource")
-            .impl(() -> new JavaProvideSourceDirTask(CompileTarget.MAIN))
+            .impl(() -> new JavaProvideSourceDirModuleTask(CompileTarget.MAIN))
             .provides("source", true)
             .desc("Provides main sources for other products.")
             .register();
 
         task("provideTestSource")
-            .impl(() -> new JavaProvideSourceDirTask(CompileTarget.TEST))
+            .impl(() -> new JavaProvideSourceDirModuleTask(CompileTarget.TEST))
             .provides("testSource", true)
             .desc("Provides test sources for other products.")
             .register();
 
         task("compileJava")
-            .impl(() -> new JavaCompileTask(buildConfig, runtimeConfiguration, CompileTarget.MAIN,
-                getRepositoryPath()))
+            .impl(() -> new JavaCompileModuleTask(CompileTarget.MAIN, getRepositoryPath()))
             .provides("compilation")
             .uses("source", "compileDependencies")
+            .importFromModules("compilation")
             .desc("Compiles main sources.")
             .register();
 
         task("compileTestJava")
-            .impl(() -> new JavaCompileTask(buildConfig, runtimeConfiguration, CompileTarget.TEST,
-                getRepositoryPath()))
+            .impl(() -> new JavaCompileModuleTask(CompileTarget.TEST, getRepositoryPath()))
             .provides("testCompilation")
             .uses("compilation", "testSource", "testDependencies")
+            .importFromModules("compilation")
             .desc("Compiles test sources.")
             .register();
 
         task("assembleJar")
-            .impl(() -> new JavaAssembleTask(buildConfig, getPluginSettings()))
+            .impl(() -> new JavaAssembleModuleTask(getPluginSettings()))
             .provides("jar")
             .uses("processedResources", "compilation")
             .desc("Assembles .jar file from compiled classes.")
             .register();
 
         task("assembleSourcesJar")
-            .impl(() -> new JavaAssembleSourcesJarTask(buildConfig))
+            .impl(JavaAssembleSourcesJarModuleTask::new)
             .provides("sourcesJar")
             .uses("source", "resources")
             .desc("Assembles .jar file from main sources and main resources.")
             .register();
 
         task("provideResources")
-            .impl(() -> new JavaProvideResourcesDirTask(CompileTarget.MAIN))
+            .impl(() -> new JavaProvideResourcesDirModuleTask(CompileTarget.MAIN))
             .provides("resources", true)
             .desc("Provides main resources for other products.")
             .register();
 
         task("provideTestResources")
-            .impl(() -> new JavaProvideResourcesDirTask(CompileTarget.TEST))
+            .impl(() -> new JavaProvideResourcesDirModuleTask(CompileTarget.TEST))
             .provides("testResources", true)
             .desc("Provides test resources for other products.")
             .register();
 
         task("processResources")
-            .impl(() -> new ResourcesTask(runtimeConfiguration, buildConfig, getPluginSettings(),
+            .impl(() -> new ResourcesModuleTask(getPluginSettings(),
                 CompileTarget.MAIN, getRepositoryPath()))
             .provides("processedResources")
             .uses("resources")
@@ -96,7 +91,7 @@ public class JavaPlugin extends AbstractPlugin<JavaPluginSettings> {
             .register();
 
         task("processTestResources")
-            .impl(() -> new ResourcesTask(runtimeConfiguration, buildConfig, getPluginSettings(),
+            .impl(() -> new ResourcesModuleTask(getPluginSettings(),
                 CompileTarget.TEST, getRepositoryPath()))
             .provides("processedTestResources")
             .uses("testResources")
@@ -104,14 +99,14 @@ public class JavaPlugin extends AbstractPlugin<JavaPluginSettings> {
             .register();
 
         task("javadoc")
-            .impl(JavadocTask::new)
+            .impl(JavadocModuleTask::new)
             .provides("javadoc")
             .uses("source", "compileDependencies")
             .desc("Creates Javadoc pages.")
             .register();
 
         task("assembleJavadocJar")
-            .impl(() -> new JavaAssembleJavadocJarTask(buildConfig))
+            .impl(JavaAssembleJavadocJarModuleTask::new)
             .provides("javadocJar")
             .uses("javadoc")
             .desc("Assembles .jar file from Javadocs.")
