@@ -21,14 +21,28 @@ import builders.loom.api.RuntimeConfiguration;
 public class RuntimeConfigurationImpl implements RuntimeConfiguration {
 
     private final boolean cacheEnabled;
+    private final String version;
+    private final boolean moduleBuild;
 
-    public RuntimeConfigurationImpl(final boolean cacheEnabled) {
+    public RuntimeConfigurationImpl(final boolean cacheEnabled, final String version,
+                                    final boolean moduleBuild) {
         this.cacheEnabled = cacheEnabled;
+        this.version = version;
+        this.moduleBuild = moduleBuild;
     }
 
     @Override
     public boolean isCacheEnabled() {
         return cacheEnabled;
+    }
+
+    @Override
+    public String getVersion() {
+        return version;
+    }
+
+    public boolean isModuleBuild() {
+        return moduleBuild;
     }
 
 }
